Enforce ~consistent~ either hash syntax (#309)
* Enforce consistent hash syntax

`EnforcedShorthandSyntax: always` is the default setting in Rubocop even though Bozhidar himself [doesn't like the option](https://batsov.com/articles/2022/01/20/bad-ruby-hash-value-omission/). I suggest to change the default option to `consistent` because I can see the advantage of the shorthand syntax when passing on options to deeper levels. But in any other context I found myself to be confused by it.

This change allows hashes to be in shorthand syntax only if the whole hash uses it consistently. See https://github.com/rubocop/rubocop/pull/10776.

* Switch to either

Co-authored-by: Alessandro Rodi <alessandro.rodi@renuo.ch>